backend/foodgram/requirements.txt
django==3.2.0
djangorestframework==3.12.4
djoser==2.1.0
djangorestframework-simplejwt==4.8.0
webcolors
Pillow
drf-extra-fields
django-filter
reportlab